Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -3,7 +3,7 @@ int main() {
     int a,b,c;
     printf("");
     scanf("%d %d %d",&a,&b,&c);
-    float x=(a+b+c)/3;
+    int x=(a+b+c)/3;
     printf("Average: %f",x);
     return 0;
 }
